requirements.txt
pandas
numpy
openpyxl
requests
tqdm
faiss-cpu
streamlit
ollama

